backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
redis==6.4.0
httpx==0.28.1
boto3==1.40.29
python-multipart==0.0.20
email-validator==2.3.0
